Gun Rework
Rebuilt the gun system from the ground up. It now has a semi-comprehensive set of values to tweak to your own desire, and should be easily modular for various weapons or upgrades. Have yet to implement on enemies yet, but should be simple enough and will be done promptly.
Removed some needless debug statements.
Improved hitboxes to make aiming less annoyin.

+++ Assets/Scripts/Damageable.cs
@@ -11,7 +11,7 @@ public class Damageable : MonoBehaviour
     public int maxHealth;
     public AnimationCurve heathOpacityCurve;
     
-    private int _health;
+    private float _health;
     private GameObject _healthBar;
 
     private float _barVisibility = 0;
@@ -47,7 +47,7 @@ public class Damageable : MonoBehaviour
         Destroy(gameObject);
     }
     
-    public void Damage(int damage)
+    public void Damage(float damage)
     {
         _barVisibility = 1;
         _health -= damage;
